Fix model loading and analysis printing bugs
1. **Fix dtype parameter error (core.py)**
   - Changed `dtype=` to `torch_dtype=` in AutoModelForCausalLM.from_pretrained()
   - Fixes: Qwen3ForCausalLM.__init__() got unexpected keyword argument 'dtype'
   - torch_dtype is the correct Hugging Face parameter name

2. **Fix NoneType format error (analysis.py)**
   - Added None check for first_refusal_alpha before formatting
   - Fixes: TypeError when no refusals detected (all responses compliant)
   - Now shows helpful message when no refusals found

These fixes allow analyze-alpha to complete successfully with style vectors
(emoji, format) where refusal detection doesn't apply.

+++ latent_control/core.py
@@ -140,7 +140,7 @@ class VectorTrainer:
         self.model = AutoModelForCausalLM.from_pretrained(
             self.config.model_path,
             quantization_config=quantization_config,
-            dtype=self.config.dtype,
+            torch_dtype=self.config.dtype,
             trust_remote_code=self.config.trust_remote_code,
             device_map="auto",
         )
